+++ CHANGELOG.md
@@ -7,6 +7,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+## [0.3.2] - 2022-02-14
+
 ### Fixed
 
 - Unquoted attribute values are now correctly parsed.
@@ -75,7 +77,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 - Initial release.
 
-[unreleased]: https://github.com/malobre/rust-vue-sfc/compare/v0.3.1...HEAD
+[unreleased]: https://github.com/malobre/rust-vue-sfc/compare/v0.3.2...HEAD
+[0.3.2]: https://github.com/malobre/rust-vue-sfc/compare/v0.3.1...v0.3.2
 [0.3.1]: https://github.com/malobre/rust-vue-sfc/compare/v0.3.0...v0.3.1
 [0.3.0]: https://github.com/malobre/rust-vue-sfc/compare/v0.2.0...v0.3.0
 [0.2.0]: https://github.com/malobre/rust-vue-sfc/compare/v0.1.0...v0.2.0